prevent CPAN::Reporter::Config from ignoring ~ on Win32
The previous behavior of C::R::C was to switch to "~/My Documents" on windows and not even try to look in ~ itself. This can be very confusing when migrating from the previous mail-based system to the new metabase system. This commit changes the behavior to default to ~ even on windows and only look in "~/My Documents" if there is no reporter config dir in ~.