fix inc module spaces in a cross-platform manner
From the message of the previous commit that changed this line i'm assuming that quotemeta was put in place to escape spaces in directory names. Thus works great on *nix systems, where \s get interpreted as escaping.

On Windows however the shell interprets \s as directory separators, which makes the resulting path completely incomprehensible to it.

"system LIST" will apply OS-specific escaping to the list elements, thus enabling it to work correctly everywhere.