Change how PrereqCheck tests if modules can load
Instead of loading via "use", we load via "require" and
the actual file path.  This is necessary to avoid
loading the wrong inc::Module::Install in certain cases
(RT #67417) as a result of the name substitution
used to fix RT #51257.

Because M::I can't have it both ways, if Module::Install
appears to be installed and appears to be older than 0.95,
it will not actually be loaded and will assumed to be
installed correctly.