Expand ~ to home directory using File::HomeDir instead of File::Glob
Using File::Glob::bsd_glob to expand the ~ to the home directory
seems easy, but as File::Glob itself admits, it doesn't work on
very many systems. Notably it does not work on Windows by default
(it works if HOME is set in the environment, but HOME it not set
by default in Windows).

This changes ~ expansion to use File::HomeDir which is much more
portable. It also changes _normalize_id_file to normalize the
directories to that of the system as well.