PrereqCheck: don't load modules with absolute path
Loading an absolute path doesn't record the module load properly
in %INC, which can cause a module to be reloaded, which occasionally
is fatal and leads modules to be reported as 'broken'.

With a relative path ("Foo/Bar.pm"), we also strip "." from @INC
to avoid picking up "t/..." or "inc/..." and thinking they
are installed.

Fixes rt.cpan.org #75559 and 76394