Record whether perl was built with taint support enabled or not
Now that perl can actually pass its own tests with taint support disabled,
CPAN::Reporter needs to pay attention to this so that we can test against the
same version of perl with and without this feature. Without this, if you have
already tested a dist against a normal perl 5.37.11 (for example) and then try
to test against 5.37.11 with no taint support, you'll not send test reports
reliably using the second build.

There are no doubt other obscure options that should be dealt with similarly,
and can be added later.